Move the handling of tests for both directions to lou_checkyaml
That way we can check the individual results for either the forward or
backward translation. Ultimately (when supported in the YAML reader)
this will allow us mark tests with xfail where only forward or
backward translation fails.

This commit just lays the ground work by doing the forward and
backward tests individually.

Of course now a lot of tests have unexpected passes where only one of
the two directions fail and the other one passes.

The next step would be to enable support for `xfail-forward` and
`xfail-backward` (or something along that line) in the YAML reader and
then "fixing" the tests.

Probably documentation needs to be adapted too.

It appears that the test suite is slower now. I suspect that could be
because lou_checkyaml is the driver now and more tests data is copied
into the test suite. But that is just a hunch.